Bugfixes for new Pandas behavior
Pandas v0.23 is more cautious about merges and was crashing when trying
to merge the data_index trial_id column to the trial_index. This is
because the index of the trial_index was 'int64', while the data_index
column was of type 'object'. Fixed by manually converting the trial_id
column to be int.